requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.15
filelock==3.0.12
Flask==1.1.2
future==0.18.2
googlemaps==4.4.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.1.1
importlib-resources==3.3.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
line-bot-sdk==1.17.0
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
pipenv==2020.11.15
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
static3==0.7.0
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
virtualenv==20.2.1
virtualenv-clone==0.5.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
